refactor(openai): split core.ts into prompt-repository + with-prompt + pure core (#67)
Decouple AI dispatch from database access by moving the supabase-touching
prompt loaders out of `src/lib/openai/core.ts` into their own module. The
sliminmed `core.ts` is now a pure AI dispatcher with zero supabase
import; consumers that only need `callWithStructuredPrompt` can import
it directly from '@/lib/openai/core' to avoid the supabase + prompt-repo
transitive imports the barrel pulls in.

New files:
- prompt-repository.ts (~190 lines) — getPrompt, getPromptJSON,
  getPromptWithProvider (the only place in this module group that
  imports supabaseAdmin)
- with-prompt.ts (~44 lines) — callAIWithPrompt: the composition layer
  that loads a prompt from the repo then calls the pure dispatcher

Slimmed:
- core.ts (391 → 156 lines): only callWithStructuredPrompt remains;
  no supabase import; imports only clients, types, response-parser

Import-site updates (1-line each):
- index.ts barrel re-sources callAIWithPrompt from ./with-prompt;
  public surface unchanged
- ai-call.ts imports callAIWithPrompt from ./with-prompt
- prompt-loaders.ts imports getPrompt from ./prompt-repository
- 3 monitoring files (alert-triage, log-analyzer, feed-health-analyzer)
  switch their direct sub-path import to './with-prompt'
- core.test.ts splits its imports across the 3 new locations
- alert-triage.test.ts updates its vi.mock target

No behavior change. 630/630 tests pass. Coverage gate (#65) clean.

Caller migration to direct sub-path imports for tree-shaking is
deliberately deferred — open a follow-up issue to migrate the 25+
barrel consumers per a per-file plan.
